requirements.txt
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
httplib2==0.14.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
requests==2.22.0
requests-unixsocket==0.2.0
SQLAlchemy==1.4.12
flask_googlemaps==1.16.0
